chore: split interfaces and implementations for player and player view
