+++ src/components/PlayerView/native.ts
@@ -1,6 +1,6 @@
 import { requireNativeComponent } from 'react-native';
 import { NativePlayerViewEvents } from './events';
-import { BasePlayerViewProps } from './index';
+import { BasePlayerViewProps } from './properties';
 import { FullscreenHandlerBridge } from '../../ui/fullscreenhandlerbridge';
 import { CustomMessageHandlerBridge } from '../../ui/custommessagehandlerbridge';
