+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './PlayerView';
 export * from './PlayerView/pictureInPictureConfig';
+export * from './PlayerView/properties';
