+++ src/index.ts
@@ -19,3 +19,5 @@ export * from './audioTrack';
 export * from './media';
 export * from './tweaksConfig';
 export * from './bufferConfig';
+export * from './playbackConfig';
+export * from './playerConfig';
